Move `beta` multn in `f(x)` and `g(x)` polynomials
This has no affect on existing assertions, but it's the form that I've
seen more commonly after seeing a few more online resources compared to
what I did previously. Furthermore, this change appears to make a
difference in round five, when the patterns in the polynomials `f(x)`
and `g(x)` appear in the `r(x)` polynomial.